Create mountedtotals
My comments in this script- if you open this up and investigate - explain everything. I have written a few different scripts trying to achieve what this does. Some of them have helped. Other ones have proven somewhat extraneous. 

I'd say - hmm, maybe expect a modification or revision to this. But as it stands I like this a lot. This will help both a desktop user running Linux at home and an admin - though it would be cool to beef this up to where it does what it does not but over a large WAN. 

If a person is a hardcore data user, data hoarder, media lover - whatever - that person quickly loses track of where any given drive is mounted, how files are interspersed, along with why any given thing is so. This script doesn't answer that question necessarily but it is a good starting point for someone wanting to investigate where their data is.  

If you're new to Linux and don't understand how the file system works, this one script isn't going to explain that for you but it might go a long way in helping. Read my comment about pseudo files and investigate the list of objects under /dev/mapper. Understanding that location will explain in part why folks say - probably a lot of these people being former Windows admins or power users - "everything in Linux is a file." That's not necessarily true from a philosophical standpoint, but it is true in reference to pseudo files. There's a lot to say about this but anyway - hope this helps a few admins get stuff done more easily.