Create installprogram
Please see the comments. I am not finished with this. I found this script somehow buried deep into my list of random shell scripts and realized I had completely forgotten I even wrote it ... so here it is. 

This works best if you know the full name of your program. There was a thought I had just now that you could implement the find command to grep for strings. The problem with that attempt is that you will then end up trying to install several programs which have nothing to do with the one you want - but are related.